update crud + add open close on installation
