+++ backend/app/src/crud/patients.py
@@ -6,15 +6,6 @@ from sqlalchemy.orm import Session
 
 from ..core.const import ADMIN_USERNAME, SALT_HASH, SMS_PATIENT_CREATE
 from ..core.excp import BadValues, DuplicateCF
-from ..core.status import (
-    INSTALLATION_CLOSED,
-    INSTALLATION_CLOSING,
-    INSTALLATION_OPEN,
-    INSTALLATION_OPENING,
-    INSTALLATION_PAUSE,
-    INSTALLATION_UNKNOW,
-    TICKET_CLOSED,
-)
 from ..ormodels import (
     Patient,
     PatientDetail,
@@ -23,11 +14,17 @@ from ..ormodels import (
     PatientScreening,
 )
 from ..schemas.contact import ContactEntry
-from ..schemas.patient import PatientCreate, PatientRead, PatientStatus, PatientUpdate
+from ..schemas.patient import (
+    PatientCreate,
+    PatientRead,
+    PatientStatus,
+    PatientUpdate,
+    PatientInfo,
+)
 from ..schemas.ticket import TicketCreate
 from ..schemas.ticket_message import TicketMessageCreate
 from ..utils import to_age, to_city
-from . import patient_contacts, tickets
+from . import patient_contacts, tickets, installation_details
 
 
 def query_one(db: Session, patient_id: int) -> PatientFull:
@@ -78,7 +75,7 @@ def read_many(db: Session) -> list[PatientStatus]:
                 result_orm.screenings[-1].neuro_diag if result_orm.screenings else None
             ),
             patient_id=result_orm.patient_id,
-            status=status(db, patient_id=result_orm.patient_id),
+            status=installation_details.status(db, patient_id=result_orm.patient_id),
             hue=arlecchino.draw(result_orm.patient_id, SALT_HASH),
         )
         for result_orm in results_orm
@@ -185,26 +182,14 @@ def update(db: Session, patient_id: int, patient: PatientUpdate) -> PatientRead:
     return result
 
 
-def status(db: Session, patient_id: int) -> str:
-    result_orm = query_one(db, patient_id)
-    ticket_status = all(
-        e.status == TICKET_CLOSED for e in tickets.read_many(db, patient_id=patient_id)
+def info(db: Session, patient_id: int) -> PatientInfo:
+    result_orm = (
+        db.query(PatientDetail).where(PatientDetail.patient_id == patient_id).one()
     )
-    context = (
-        result_orm.date_start is not None,
-        result_orm.date_end is not None,
-        ticket_status,
+    result = PatientInfo(
+        first_name=result_orm.first_name,
+        last_name=result_orm.last_name,
+        home_address=result_orm.home_address,
+        contacts=patient_contacts.read_many(db, patient_id),
     )
-    match context:
-        case (True, False, True):
-            return INSTALLATION_OPEN
-        case (True, False, False):
-            return INSTALLATION_PAUSE
-        case (_, True, True):
-            return INSTALLATION_CLOSED
-        case (False, _, False):
-            return INSTALLATION_OPENING
-        case (True, True, False):
-            return INSTALLATION_CLOSING
-        case _:
-            return INSTALLATION_UNKNOW
+    return result
